Canvas部分をusePongGameに変更
